♻️ use usize instead of u32 indexable dimensions

+++ src/core.rs
@@ -47,26 +47,26 @@ impl Display for OpenErr {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Game {
     state: GameState,
-    width: u32,
-    height: u32,
-    num_mines: u32,
-    randomizer: fn(u32, u32, u32, u32, u32) -> Vec<Vec<bool>>,
-    clear_remaining: u32,
+    width: usize,
+    height: usize,
+    num_mines: usize,
+    randomizer: fn(usize, usize, usize, usize, usize) -> Vec<Vec<bool>>,
+    clear_remaining: usize,
     mines: Vec<Vec<bool>>,
     opened: Vec<Vec<bool>>,
     neighbours: Vec<Vec<u8>>,
 }
 
 impl Game {
-    pub fn new(width: u32, height: u32, num_mines: u32) -> Self {
+    pub fn new(width: usize, height: usize, num_mines: usize) -> Self {
         Self::new_with(height, width, num_mines, default_randomizer)
     }
 
     pub fn new_with(
-        width: u32,
-        height: u32,
-        num_mines: u32,
-        randomizer: fn(u32, u32, u32, u32, u32) -> Vec<Vec<bool>>,
+        width: usize,
+        height: usize,
+        num_mines: usize,
+        randomizer: fn(usize, usize, usize, usize, usize) -> Vec<Vec<bool>>,
     ) -> Self {
         if width < 1 || height < 1 {
             panic!("Invalid size");
@@ -87,15 +87,15 @@ impl Game {
         }
     }
 
-    pub fn width(&self) -> u32 {
+    pub fn width(&self) -> usize {
         self.width
     }
 
-    pub fn height(&self) -> u32 {
+    pub fn height(&self) -> usize {
         self.height
     }
 
-    pub fn num_mines(&self) -> u32 {
+    pub fn num_mines(&self) -> usize {
         self.num_mines
     }
 
@@ -106,14 +106,14 @@ impl Game {
         self.neighbours = Vec::new();
     }
 
-    pub fn reset(&mut self, width: u32, height: u32, num_mines: u32) {
+    pub fn reset(&mut self, width: usize, height: usize, num_mines: usize) {
         self.height = height;
         self.width = width;
         self.num_mines = num_mines;
         self.clear();
     }
 
-    pub fn reset_randomizer(&mut self, randomizer: fn(u32, u32, u32, u32, u32) -> Vec<Vec<bool>>) {
+    pub fn reset_randomizer(&mut self, randomizer: fn(usize, usize, usize, usize, usize) -> Vec<Vec<bool>>) {
         self.randomizer = randomizer;
     }
 
@@ -121,15 +121,15 @@ impl Game {
         self.state
     }
 
-    pub fn get_cell(&self, x: u32, y: u32) -> Cell {
+    pub fn get_cell(&self, x: usize, y: usize) -> Cell {
         match self.state {
             GameState::New => Cell::Hidden,
-            GameState::Playing if !self.opened[y as usize][x as usize] => Cell::Hidden,
+            GameState::Playing if !self.opened[y][x] => Cell::Hidden,
             _ => {
-                if self.mines[y as usize][x as usize] {
+                if self.mines[y][x] {
                     Cell::Mine
                 } else {
-                    Cell::Clear(self.neighbours[y as usize][x as usize])
+                    Cell::Clear(self.neighbours[y][x])
                 }
             }
         }
@@ -141,7 +141,7 @@ impl Game {
             .collect()
     }
 
-    pub fn open(&mut self, x: u32, y: u32) -> Result<OpenInfo, OpenErr> {
+    pub fn open(&mut self, x: usize, y: usize) -> Result<OpenInfo, OpenErr> {
         if x >= self.width || y >= self.height {
             return Err(OpenErr::OutOfBounds);
         }
@@ -152,11 +152,11 @@ impl Game {
                 self.open(x, y)
             }
             GameState::Playing => {
-                if self.opened[y as usize][x as usize] {
+                if self.opened[y][x] {
                     Err(OpenErr::AlreadyOpen)
                 } else {
-                    self.opened[y as usize][x as usize] = true;
-                    if self.mines[y as usize][x as usize] {
+                    self.opened[y][x] = true;
+                    if self.mines[y][x] {
                         self.state = GameState::Lost;
                         Ok(OpenInfo {
                             state: self.state,
@@ -169,7 +169,7 @@ impl Game {
                         }
                         Ok(OpenInfo {
                             state: self.state,
-                            cell: Cell::Clear(self.neighbours[y as usize][x as usize]),
+                            cell: Cell::Clear(self.neighbours[y][x]),
                         })
                     }
                 }
@@ -180,20 +180,20 @@ impl Game {
 
     // Utility functions
 
-    fn generate_board(&mut self, firstx: u32, firsty: u32) {
+    fn generate_board(&mut self, firstx: usize, firsty: usize) {
         self.mines = (self.randomizer)(self.width, self.height, self.num_mines, firstx, firsty);
-        self.opened = vec![vec![false; self.width as usize]; self.height as usize];
+        self.opened = vec![vec![false; self.width]; self.height];
         self.calculate_neighbours();
         self.clear_remaining = self.width * self.height - self.num_mines;
     }
 
     fn calculate_neighbours(&mut self) {
-        self.neighbours = vec![vec![0; self.width as usize]; self.height as usize];
+        self.neighbours = vec![vec![0; self.width]; self.height];
         for y in 0..self.height {
             for x in 0..self.width {
-                if self.mines[y as usize][x as usize] {
-                    for j in max(1, y as usize) - 1..min(self.height, y + 2) as usize {
-                        for i in max(1, x as usize) - 1..min(self.width, x + 2) as usize {
+                if self.mines[y][x] {
+                    for j in max(1, y) - 1..min(self.height, y + 2) {
+                        for i in max(1, x) - 1..min(self.width, x + 2) {
                             self.neighbours[j][i] += 1;
                         }
                     }
@@ -204,17 +204,17 @@ impl Game {
 }
 
 fn default_randomizer(
-    width: u32,
-    height: u32,
-    num_mines: u32,
-    firstx: u32,
-    firsty: u32,
+    width: usize,
+    height: usize,
+    num_mines: usize,
+    firstx: usize,
+    firsty: usize,
 ) -> Vec<Vec<bool>> {
-    let mut mines = vec![vec![false; width as usize]; height as usize];
+    let mut mines = vec![vec![false; width]; height];
     let mut mines_left = num_mines;
     let area = width * height;
     let spaces = area - num_mines - 1;
-    mines[firsty as usize][firstx as usize] = true;
+    mines[firsty][firstx] = true;
 
     let mut rng = rand::thread_rng();
     while mines_left > 0 {
@@ -222,18 +222,18 @@ fn default_randomizer(
         let mut i = 0;
         for j in 0..area {
             let (x, y) = (j % width, j / width);
-            if mines[y as usize][x as usize] {
+            if mines[y][x] {
                 continue;
             }
             if i == r {
-                mines[y as usize][x as usize] = true;
+                mines[y][x] = true;
                 mines_left -= 1;
                 break;
             }
             i += 1;
         }
     }
-    mines[firsty as usize][firstx as usize] = false;
+    mines[firsty][firstx] = false;
     mines
 }
 
@@ -242,17 +242,17 @@ mod tests {
     use super::*;
 
     fn dummy_randomizer(
-        width: u32,
-        height: u32,
-        num_mines: u32,
-        _firstx: u32,
-        _firsty: u32,
+        width: usize,
+        height: usize,
+        num_mines: usize,
+        _firstx: usize,
+        _firsty: usize,
     ) -> Vec<Vec<bool>> {
-        let mut mines = vec![vec![false; width as usize]; height as usize];
+        let mut mines = vec![vec![false; width]; height];
         for i in 0..num_mines {
             let x = i % width;
             let y = i / width;
-            mines[y as usize][x as usize] = true;
+            mines[y][x] = true;
         }
         mines
     }
@@ -359,7 +359,7 @@ mod tests {
     fn randomizer_does_not_get_first() {
         for i in 0..5 {
             let mines = default_randomizer(5, 5, 24, i, i);
-            assert_eq!(mines[i as usize][i as usize], false);
+            assert_eq!(mines[i][i], false);
         }
     }
 }
